Dialogs: -Wifi -GPS Refactoring

+++ app/src/main/java/com/gmartinsribeiro/weather/android/entity/Rain.java
@@ -10,14 +10,14 @@ public class Rain {
 
     @SerializedName("3h")
     @Expose
-    private Integer _3h;
+    private Double _3h;
 
     /**
      *
      * @return
      * The _3h
      */
-    public Integer get3h() {
+    public Double get3h() {
         return _3h;
     }
 
@@ -26,7 +26,7 @@ public class Rain {
      * @param _3h
      * The 3h
      */
-    public void set3h(Integer _3h) {
+    public void set3h(Double _3h) {
         this._3h = _3h;
     }
 }
